Fixed cast.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@2080 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/fractalpooluser.cc
@@ -303,7 +303,7 @@ void FractalPU::startNextJob()
                                     xStep, yStep, color.rgb());
                }
                char str[64];
-               snprintf((char*)&str, sizeof(str), "%u", number + 1);
+               snprintf((char*)&str, sizeof(str), "%u", (unsigned int)number + 1);
                const QString numberString(str);
                painter.drawText(xPosition * xStep, yPosition * yStep, xStep, yStep, Qt::AlignCenter, numberString);
             }
